fix: make externally-referenced IDs AUTOINCREMENT
This forces SQLite not to reuse the Row ID. This could be a problem if a lecturer were deleted & re-created. Then, the department widget wouldn't realise that anything changed as their ID would still be the same.

This is also the case with content_streams - if a content stream were deleted & recreated then the displays referring to that content would start listening to the new stream which is not what we want

Signed-off-by: Restioson <restiosondev@gmail.com>